Refactor Contacts Header to use it in Settings and next Apps
